refactor: promexp.py file, logger
