+++ promexp/provider.py
@@ -1,16 +1,16 @@
+from promexp.logger import logger
+
+
 class BaseProvider:
     name = "base"
 
     def __init__(self, parent, settings):
         self.parent = parent
-        self.logger = parent.logger
         self.enabled = settings is not None
         if isinstance(settings, dict):
             self.settings = settings
         else:
-            self.logger.warning(
-                f"Incorrect settings for {self.name} provider. Ignoring"
-            )
+            logger.warning(f"Incorrect settings for {self.name} provider. Ignoring")
             self.settings = {}
 
     def __getitem__(self, key):
